refactor: rename handler and repository files, update import paths for agent-related modules
