+++ apps/electron/src/main/modules/agent/deployed-agent-repository.ts
@@ -1,6 +1,6 @@
 import { DeployedAgent } from "@mcp_router/shared";
-import { BaseRepository } from "../../core/base-repository";
-import { SqliteManager } from "../../core/sqlite-manager";
+import { BaseRepository } from "../../infrastructure/database/core/base-repository";
+import { SqliteManager } from "../../infrastructure/database/core/sqlite-manager";
 import { v4 as uuidv4 } from "uuid";
 
 /**
